requirements.txt
anyio==3.3.0
APScheduler==3.7.0
arrow==1.1.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
fastapi==0.68.0
Flask==2.0.1
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
hyperframe==5.2.0
idna==3.2
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
jinja2-time==0.2.0
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
mwclient==0.10.1
nb-cli==0.5.0
nonebot-adapter-cqhttp==2.0.0a15
nonebot2==2.0.0a15
oauthlib==3.1.1
OpenCC==1.1.1.post1
Pillow==8.3.1
poyo==0.5.0
prompt-toolkit==3.0.20
pydantic==1.8.2
pyfiglet==0.8.post1
pygtrie==2.4.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
text-unidecode==1.3
tomlkit==0.7.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.1
Xdi8Translator==0.4.0
